feat: SwiftUI View Rendering Instrumentation (#20)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes #<enter issue here>

## Short description of the changes
SwiftUI rendering is very interesting! 

This PR currently has a whole bunch of instrumentation and logging which
makes it easier to see what's going on

This PR intentionally does not ship a view modifier form of the
instrumentation. This is because ViewModifiers receive a `Content`
struct (vs. Views that receive an `@ViewBuilder () -> Content`
callback!); that is: ViewModifiers receive a view _after it has been
initialized_, and therefore have no way to retroactively start a span
before the view begins initializing.

Compare the spans/log output from using the wrapper view form:


![image](https://github.com/user-attachments/assets/efc7182b-2657-4fda-b3b3-f33778e2056b)

with the view modifier form:

![image](https://github.com/user-attachments/assets/f6197366-2cea-4121-bd4b-f31b53778439)

And, now that we've settled on the wrapper form, here's the kinds of
traces we're producing:


![image](https://github.com/user-attachments/assets/f4d07cd1-8db3-46e0-8c1f-8d4befa33f1c)


## How to verify that this has the expected result