fix: Auto-select the SDK when running smoke tests. (#24)
## Which problem is this PR solving?

Currently, our smoke tests are hard-coded to a particular SDK we know is
available in circleci. But individual developers have updated Xcode and
no longer have that version. So, everyone has to edit this file to make
the smoke tests run. This PR changes the script to autodetect the SDK
version to use.

## Short description of the changes

`xcodebuild -showsdks` lists the SDKs. We filter for `iphonesimulator`,
and then use an ad-hoc regex to get just the name of the SDK.

## How to verify that this has the expected result

I tested it locally. If it still passes in circleci, it should be good
to go.