feat: instrumented NavigationStack in SwiftUI (#23)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Provides helpers for instrumentating navigations

## Short description of the changes
Exposes 2 APIs:

- A ViewModifier: `.instrumentNavigations(path: [any])` or
`.instrumentNavigations(path: NavigationPath)`
- A helper function: `reportNavigation(path: any)`

(I'm not overly attached to this interface so definitely let me know if
we think we should change it. ex. I could see us shipping a
`reportNavigationOnAppear()` view modifier to encapsulate the `onAppear
{ reportNavigation() }` pattern I have going, or we could change the
names of some of these, or whatever...)

Both emit the same types of the spans and also set the current path in
global state. This PR also sets up a SpanProcessor that reads that
global state and includes the current path as an attribute on the span.

(More info and full documentation included in the changes to
`README.md`)

## How to verify that this has the expected result
- [x] smoke tests
- [x] run the sample app, click around, watch the navigation spans come
through in the collector/stdout/your honeycomb account