feat: UIKit instrumentation now sets the current view (#26)
## Which problem is this PR solving?
UIKit Instrumentation was emitting `viewDidAppear` and
`viewDidDisappear` events, but wasn't persisting the "current screen"
value as the SwiftUI Navigation instrumentation did in #23

## Short description of the changes
UIKit Instrumentation now tracks the "current screen" value

Computing "current screen" is a little complex but tldr we use the
following (in order of precedence):
- `accessiblityIdentifier`
- Storybook Identifier
- `view.title`
- current class name of the view controller

`honeycombIdentifier` is provided as an escape hatch for clients to opt
out of this precedence chain and set some explicit value

This PR also updates our smoke test app to use nested UIKit navigations
as a way to exercise the "current screen" computations.

Note that we don't do anything in the `viewDidDisappear` event: I
considered setting the current screen to `nil` but opted against that
because
1. it seemed like there were cases where, View B is replacing View A,
View A's `viewDidDisappear` would run _after_ View B's `viewDidAppear`.
Thus setting current screen to `nil` would actually clobber the (now
correct) value corresponding to whatever we get from View B
2. I reasoned that if one view is disappearing, then it must be replaced
by some other view, and that other view is responsible to setting
current screen (either automatically, if the new view is a UIKit view,
or explicitly, if it's a SwiftUI view)

## How to verify that this has the expected result
- [x] run the smoke test app and manually inspect collector output
- [x] updated smoke tests