feat: actually expose the deterministic sampler (#31)
## Which problem is this PR solving?

Exposes sampling functionality (basically the same as
https://github.com/honeycombio/honeycomb-opentelemetry-android/pull/34)

## Short description of the changes

We already had a `sampleRate` option and the code for a deterministic
sampler, this PR just ties them together.

## How to verify that this has the expected result
🤔 

Existing smoke tests pass, so we know the default value (ie. `1`) does
not apply any sampling. I'm open to suggestions on how to test actual
sample rates.