fix: move session in notification from "object" to "userInfo" (#34)
## Which problem is this PR solving?

Before this PR, `SessionManager` was sending notifications when the
session changes with a `HoneycombSession` as the `object` of the
notification. But `object` is poorly-named, and is actually intended to
be the _sender_ of the notification. Its use is for filtering
notifications when setting up observers. But for something like a
session, there's a new session every time, so filtering for a specific
one doesn't really work. The sender should be the `SessionManager`, and
the session itself should be moved to `userInfo`.

## Short description of the changes

* Moves the session to `userInfo["session"]`, and changes `object` to be
the `SessionManager`.
* Marks `SessionManager` as private, since all of its members were
private anyway.

## How to verify that this has the expected result

The tests have been updated.

- [✅] Changelog is updated