feat: Add cocopods frameworks for publishing (#35)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes #<enter issue here>

## Short description of the changes

Adds coccoapods `podspec` file to allow us to publish this on the
cocoapods repo.

This also conditionally disables gRPC protocol if the grpc libraries are
unavailable. This was necessary as there is no official `grpc-swift`
package available on cocoapods upon which upstream otel-swift depends.

## How to verify that this has the expected result

## TODO:

- Publishing this to cocoapods (preferably as part of release)
- ~~Upstream changes merged (so that people can download the otel
dependencies).~~

---

- [X] Changelog is updated