maint: Add a .clang-format and instructions for running it. (#48)
## Which problem is this PR solving?

Our Objective-C code has no linter or auto-formatter. We want this to be
automatic so that we don't waste time on formatting discussion in code
reviews.

## Short description of the changes

* Adds instructions for installing `clang-format`.
* Adds a `.clang-format` config file based on OTel and existing code.
* Extends `make format` to include Objective-C code.
* Auto-formats the existing Objective-C code.

Note that this does _not_ extend _lint_ to cover Objective-C, as that is
non-trivial, and we can always decide to add that later.

## How to verify that this has the expected result

It's a purely cosmetic change, so existing smoke-tests should be
sufficient.

---

- [N/A] Changelog is updated