docs: trace propagation (#56)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Users will need to know what steps to take to propagate traces to their
own backend. This adds documentation on how to do that.

- Closes #<enter issue here>

## Short description of the changes
You need to create an active span in order for all events to be on the
same trace (frontend and backend) this differs
from android where you can just use the Otel Context. Swift uses the
span context to set this up.

## How to verify that this has the expected result


---

- [X] CHANGELOG is updated
- [X] README is updated with documentation