feat: add network connection attributes (#57)
## Which problem is this PR solving?

It would be nice to have attributes on spans about the type of network
connection.

## Short description of the changes

This PR uses the `NetworkStatus` package from OTel to add
network-specific attributes. OTel uses this package only in their
`URLSession` instrumentation, and they spin up a new `NetworkMonitor` on
every request.

However, in this PR, we create one `NetworkMonitor` when the Honeycomb
SDK is initialized, and we use a new custom span processor to add the
attributes to every span.

## How to verify that this has the expected result

I added a new smoke test to verify the attributes. Unfortunately, with
the smoke tests, we can only verify the "wifi" value.

---

- [X] CHANGELOG is updated
- [N/A] README is updated with documentation