maint: remove cocoapods support (#59)
## Which problem is this PR solving?

With our new dependencies on the `PersistenceExporter` and
`NetworkStatus` packages in upstream OTel, our Cocoapods support is
broken. Before we can reasonably support and maintain Cocoapods in our
repo, we need to do the upstream work to make every OTel library we
might use support Cocoapods, and all of those pods need tests to
guarantee we don't break them with future new dependencies. So, for now,
we are dropping Cocoapods support.

## Short description of the changes

This simply removes the cocoapods file.

The only other cocoapods-related change was guarding the compilation of
our GRPC code on whether GRPC is available. I think it's fine to leave
that in.

## How to verify that this has the expected result

N/A

---

- [*] CHANGELOG is updated
- [N/A] README is updated with documentation