fix: sometimes crashes dont flush (#61)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Due to the flakiness of the underlying `flush` methods, sometimes
exception logs do not get sent before the app crashes.

- Closes #<enter issue here>

## Short description of the changes

We are adding a 3 second wait to allow the crash exception to finish
exporting.

## How to verify that this has the expected result

---

- [X] CHANGELOG is updated
- [N/A] README is updated with documentation