feat: navigation instrumentation can track time spent on a screen (#64)
## Which problem is this PR solving?
We would like to able to track total time spent in a
screen/view/activity

## Short description of the changes
Instead of emitting a single `Navigation` span, we now emit
`NavigationTo` and `NavigationFrom` spans. `NavigationTo` spans are the
equivalent of the old `Navigation` spans, and the `NavigationFrom` span
is entirely new. It includes attributes for the screen we are leaving,
as well as time spent on that screen.

We also hook into UIKit app lifecycle to ensure that the timing of those
spans is correct (ie. we're not counting time spent while the phone was
locked or the app was in the background).

Navigation spans now include a `navigation.trigger` attribute, which is
either `navigation` or an application lifecycle method (ex.
`appDidEnterBackground` etc.). This attribute can also be set by clients
via an optional `reason: String` attribute on all of our navigation
helpers.

This PR also fixes the `[]` array issue by explicitly looking for empty
paths and emitting `/` instead. We also now accept an optional `prefix:
String` param that gets prepended to the path, reducing the likelihood
of empty paths and allowing clients to distinguish between different
NavigationStacks in their applications.

## How to verify that this has the expected result
- [x] smoke tests pass

---

- [x] CHANGELOG is updated
- [x] README is updated with documentation

---------

Co-authored-by: Bee Klimt <beeklimt@honeycomb.io>