feat: add build info to resource attributes (#85)
## Which problem is this PR solving?
Adds resource attributes to enable us to symbolicate stacktraces from
our exception logger

## Short description of the changes
The important ones are `app.debug.buildUUID` and `app.debug.binaryName`.
In my tests, `app.debug.binaryName` spat out
`/Library/Developer/CoreSimulator/Volumes/iOS_22C150/Library/Developer/CoreSimulator/Profiles/Runtimes/iOS
18.2.simruntime/Contents/Resources/RuntimeRoot/usr/lib/libLogRedirect.dylib`,
not the actual executable name like I expected. This might be because I
was running the app in the xcode simulator; we'll have to experiment
with an actual compiled app on a device.

Anyways I added `app.bundle.executable` to make up for that, which looks
like it contains the actual binary name in it.

While I was in here, I added code to pull the version number info out of
the bundle as well, since that was fairly straightforward.

## How to verify that this has the expected result
- [x] Tests pass

---

- [x] CHANGELOG is updated
- [x] README is updated with documentation