feat: make HoneycombOptions builder object ObjectiveC friendly (#97)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

To support react native we need to allow access to the options builder
to ObjectiveC.

- Closes #<enter issue here>

## Short description of the changes
Make builder inherit `NSObject` and give it and applicable methods the
`@objc` directive.

## How to verify that this has the expected result

---

- [X] CHANGELOG is updated
- [N/A] README is updated with documentation