fix: make builder class open (#100)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

We need to be able to expose the builder class to React Native.

- Closes #<enter issue here>

## Short description of the changes

Make builder class `open` and enable us to extend it outside of swift.

## How to verify that this has the expected result

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation