fix: make HoneycombNavigationPathSpanProcessor threadsafe (#113)
## Which problem is this PR solving?

We've had reports of crashes in the
`HoneycombNavigationPathSpanProcessor`, and it looks like the handling
of the current path was not threadsafe, so this PR makes it threadsafe.

## Short description of the changes

We're just guarding each access of the current path members with an
`NSLock`, similar to what we did in the Session Manager.

## How to verify that this has the expected result

The smoke tests still pass.

---

- [X] CHANGELOG is updated
~- [ ] README is updated with documentation~ N/A